Remove pretty encoding from showLogMessage
While it's nice for humans to see pretty log messages and all those
identation it's much more difficult for machine consumption.

This introduces a change from using `encodePretty'` to simple `encode`
which will produce jsonline for each log message.

In modern word of containers and K8s it is the required way for logging.