Add separate jsonline logger
In order to keep backwards compatibility with previous versions of the
library for conveniency it was added another logger which fully
dedicated to jsonline logging.
This separate logger dumps the logs to stdout and can be setup with `withSimpleJsonLineStdOutLogger` which
will take care of converting the logs into jsonline, writing to stdout
and flushing the buffers.